Added authentication to deleteAccount

+++ controllers/accountController.js
@@ -48,6 +48,16 @@ const deleteAccount = async (req, res) => {
     try {
         const {Token, _id} = req.body
 
+        const user = await User.findOne({Token: Token})
+
+        if(!user) {
+            return res.status(401).json({ error: 'Invalid token' })
+        }
+
+        if(user.ValidUntil < Date.now()) {
+            return res.status(401).json({ error: 'Token expired' })
+        }
+
         const account = await Account.findByIdAndDelete(_id)
 
         return res.status(200).json({ account })
